% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_nasa_power_point.R
\name{add_efficiency}
\alias{add_efficiency}
\title{Add efficiency data to solar panel data}
\usage{
add_efficiency(panel_csv)
}
\arguments{
\item{panel_csv}{solar panel data as a .csv}
}
\value{
a tibble with calculated efficiency for each panel entry
}
\description{
Calculate panel efficiency based on solar irradience from NASA POWER API
}
\examples{
\dontrun{
panel_data = add_efficiency("adjusted_dates.csv")
}
}
