% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reliability.R
\name{convert_excel_date}
\alias{convert_excel_date}
\title{Excel date conversion helper}
\usage{
convert_excel_date(excel_date)
}
\arguments{
\item{date}{Numeric. Date in Excel serial format.}
}
\value{
A date object in YYYY-MM-DD format
}
\description{
Convert Excel serial dates to R Date format
}
