% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indices.R
\name{getPpks_farm}
\alias{getPpks_farm}
\title{Get Process Performance Index for Solar Farm}
\usage{
getPpks_farm(panel_data, power_spec)
}
\arguments{
\item{panel_data}{a tibble, as outputted from add_efficiency()}

\item{power_spec}{numeric, the per solar panel power output specification 
for the solar farm}
}
\value{
a tibble containing the calculated Ppk value for each solar far based on
a per solar panel power output specification
}
\description{
Calculates the process performance index for each solar farm based on a
specified power output specification
}
\details{
- Ppk is used to evaluate process performance relative to specification limits. 
Here, the specification limit used is provided by the user as power_spec
}
\examples{
indices_f = getPpks_farm(panel_data, 0.4)

}
