% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indices.R
\name{getPpks_panel}
\alias{getPpks_panel}
\title{Get Process Performance Index for each Solar Panel}
\usage{
getPpks_panel(panel_data)
}
\arguments{
\item{panel_data}{a tibble, as outputted from add_efficiency()}
}
\value{
a tibble containing the calculated Ppk value for each solar panel
}
\description{
Calculates the process performance index for each solar panel based on the 
nominal efficiency for each solar panel type (M, P, or T)
}
\details{
- Ppk is used to evaluate process performance relative to specification limits. 
Here, the specification limit used is the nominal efficiency for each solar panel time
}
\examples{
indices_p = getPpks_panel(panel_data)

}
