% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indices.R
\name{ppk}
\alias{ppk}
\title{Process Performance Index (for skewed, uncentered data)}
\usage{
ppk(mu, sigma_t, lower = NULL, upper = NULL)
}
\arguments{
\item{mu}{Numeric. Process mean.}

\item{sigma_t}{Numeric. Process standard deviation.}

\item{lower}{Numeric of NULL. Lower specification limit.}

\item{upper}{Numeric or NULL. Upper specification limit.}
}
\value{
Numeric. The calculated Ppk value.
}
\description{
Calculates the process performance index given a process mean, standard deviation, 
 and specification limits.
}
\details{
- Ppk is used to evaluate process performance relative to specification limits
}
\examples{
ppk(0.18, 0.02, lower = 0.17, upper = 0.20)
ppk(0.18, 0.02, upper = 0.20)
ppk(0.18, 0.02, lower = 0.17)
}
