% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{spc_all}
\alias{spc_all}
\title{Create an average SPC chart for the max power output for each day}
\usage{
spc_all(panel_data, site_id, var)
}
\arguments{
\item{panel_data}{a tibble, as outputted from add_efficiency()}

\item{site_id}{character, the unique ID of the solar panel farm of interest}

\item{var}{character, variable of interest - "e" for efficiency, "p" for power output}
}
\value{
A ggplot2 of the average SPC chart.
}
\description{
This function calculates daily peak values for a specified variable (e.g. power output)
 across one or more panels, computes the daily mean, and control limits, 
 and produces an averages chart using ggplot2. It also prints a tibble that assesses 
 whether the data passes the eight SPC checks.
}
\details{
- Daily peaks are calculated for each panel, then averaged across panels.
- Control limits are calculated as the grand mean ± 3 standard erros.
- Ribbon highlights the control limits, and daily means are shown as points and lines.
}
\examples{
\dontrun{
g1 <- spc_all(panel_data, site_id, var)
print(g1)
}
}
