% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_tests.R
\name{spcTest}
\alias{spcTest}
\title{Run SPC Tests on Subgroup}
\usage{
spcTest(averages, sigma)
}
\arguments{
\item{averages}{Numeric. Subgroup average values.}

\item{sigma}{Numeric. Standard deviation of the subgroup averages.}
}
\value{
A tibble with colums for each test. Each entry is 'TRUE' if the subgroup
 averages pass the test and 'FALSE' if they fail.
}
\description{
This function evaluates whether a vector passes each of the standard SPC tests.
 Each test checks for different types of patterns such as runs below/above control 
 limits, trends, or oscillations.
}
\details{
- Test 1: one point beyond zone a
- Test 2: two out of three points in a row beyond zone b
- Test 3: four out of five points in a row beyond zone c
- Test 4: eight points in a row on the same side of the centerline
- Test 5: six points in a rox steadily increasing or decreasing
- Test 6: fifteen points in a row in zone c
- Test 7: eight points in a rox on both sides of the centerline (none in zone c)
- Test 8: fourteen points in a row alternating up and down
}
\examples{
\dontrun{
averages <- c(10.1, 10.2, 9.9, 10.3, 10.0)
sigma <- sd(averages)
spcTest(averages, sigma)
}
Window helper function
counts the number of 'TRUE' values within a window of specified size
data, a logical vector of booleans
len, Interger. the window size
return a numeric vector where each element represents the number of 'TRUE' values
 within each rolling window
}
